fix(www): filter pending/abandoned photos from queries, user-error for 422/413, isolate Sharp module
- fetchPhotosByListingIds: add eq(status, 'complete') filter so pending/abandoned rows never produce broken public URLs
- insertPendingPhoto count: exclude abandoned rows from the cap check so failed uploads don't block future ones
- PhotoServiceError 422/413: throw UserError instead of capturing to Sentry; only 5xx hits Sentry
- PhotoServiceError message: store body as separate property, never embed in Error.message
- listing-photo-upload.server: remove top-level sharp init and Sharp pipeline; Sharp code moved to listing-photo-sharp-pipeline.server.ts so the upload route (listing-photos.ts) never triggers libvips on import
- Tests updated to import uploadListingPhoto/MAX_IMAGE_PIXELS/PUB_MAX_DIMENSION from the new pipeline file
